backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Embedding providers
openai>=1.0.0
cohere>=5.0.0
sentence-transformers>=2.2.0

# Vector search & ML
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.4

# Utilities
aiofiles>=23.0.0
